working branch

+++ src/main/java/model/BookRepository.java
@@ -4,9 +4,9 @@ import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface BookRepository extends CrudRepository <Book, Integer>{
+public interface BookRepository extends CrudRepository <Book, Long>{
 
     List<Book> findByBookName(String bookName);
 
-    Book findById(int id);
+    Book findById(long id);
 }
